Refactor: move air-quality sensor functions to utils
